requirements.txt
Flask>=2.0
requests>=2.28
python-dotenv>=1.0
flask-cors>=3.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==0.27.8
